Extract CSharpText lexer and declaration index
Move the dependency-free C# scanner, declaration index, and their existing test corpus out of the authored-source slicer. Keep the slicer on the same scanner implementation through an internal CSharpText seam.


Copilot-Session: bc2613de-a721-4efd-8390-56db684eefd7
