+++ AGENTS.md
@@ -179,6 +179,7 @@ Tests use xUnit executable projects. **Use `dotnet run`, not `dotnet test`**;
 | CLI and product output | `dotnet run --project src/dotnet-inspect.Tests -c Release` |
 | Analysis | `dotnet run --project src/ILInspector.Analysis.Tests -c Release` |
 | Decompiler | `dotnet run --project src/ILInspector.Decompiler.Tests -c Release` |
+| C# text | `dotnet run --project tests/CSharpText.Tests -c Release` |
 | Shared services | `dotnet run --project src/DotnetInspector.Services.Tests -c Release` |
 | Metadata and SourceLink | `dotnet run --project tests/ILInspector.Metadata.Tests -c Release` |
 | Metadata rendering and `mdi` | `dotnet run --project tests/DotnetInspector.MetadataRendering.Tests -c Release` |
