+++ tests/CSharpText.Tests/ConditionalRecoveryFuzzTests.cs
@@ -1,13 +1,12 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using DotnetInspector.CSharpBodySlicer;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Xunit;
 
-namespace DotnetInspector.CSharpBodySlicer.Tests;
+namespace CSharpText.Tests;
 
 // ---------------------------------------------------------------------------
 // Conditional-recovery differential fuzzer for DeclarationIndex (PR #3680).
